f32 -> f64

+++ benches/debug-simulator.rs
@@ -15,7 +15,7 @@ fn main() {
 fn controlled_id_all_control_but_one(n_qubits: usize) {
     let n_controls = n_qubits - 1;
 
-    let id_2x2 = DMatrix::<Complex<f32>>::identity(2, 2);
+    let id_2x2 = DMatrix::<Complex<f64>>::identity(2, 2);
     let mat = DebugSimulator::expand_matrix(
         id_2x2,
         &(0..n_controls).collect::<Vec<usize>>(),
@@ -23,7 +23,7 @@ fn controlled_id_all_control_but_one(n_qubits: usize) {
         n_qubits,
     );
     let dim = 1 << n_qubits;
-    let id_big = DMatrix::<Complex<f32>>::identity(dim, dim);
+    let id_big = DMatrix::<Complex<f64>>::identity(dim, dim);
 
     assert!(equal_to_matrix_c(&mat, &id_big, 0.000001,));
 }
